refactor: Add vite-plugins export
