+++ packages/kernel-browser-runtime/vite.config.ts
@@ -1,8 +1,11 @@
 // eslint-disable-next-line spaced-comment
 /// <reference types="vitest" />
 
-import { deduplicateAssets, jsTrustedPrelude } from '@ocap/repo-tools';
-import type { PreludeRecord } from '@ocap/repo-tools';
+import {
+  deduplicateAssets,
+  jsTrustedPrelude,
+} from '@ocap/repo-tools/vite-plugins';
+import type { PreludeRecord } from '@ocap/repo-tools/vite-plugins';
 import path from 'path';
 import { defineConfig } from 'vite';
 import { checker as viteChecker } from 'vite-plugin-checker';
