+++ packages/repo-tools/src/vite-plugins/deduplicate-assets.ts
