+++ packages/repo-tools/src/vite-plugins/extension-dev.ts
