+++ packages/repo-tools/src/vite-plugins/html-trusted-prelude.ts
