+++ packages/repo-tools/src/vite-plugins/index.ts
