+++ packages/repo-tools/src/vite-plugins/js-trusted-prelude.ts
